Added playlist and video url support. Still WIP, but I thought I'd update the repository.
Also fixed some issues, but a lot still remain lol.

Switched to a new scraper that supports getting metadata for video urls and playlists. It currently has a bug  where the getVideo() function returns undefined no matter the input. I've reported this bug to the author but in the meantime it can be fixed by manually editing the files in the module itself.

For this, I recommend creating a separate file to execute the .getVideo() function in yt.js (with valid input, being a YouTube video id) and going into the files mentioned in the errors. In here, remove all the square brackets after response.data (e.g: data[3] -> data) you encounter. I believe I had to change it in 5/6 files. After removing the square brackets it worked and returned the right output.